[UGLY HACK] ld: Align leading empty sections as first non-empty section
When we encounter empty sections leading a segment (such as .tls/.data
in an explicitly created bss) we need to raise the alignment of those
initial empty sections such that the following non-empty section can be
at the same location, otherwise they are placed such that the following
non-empty section is offset from them and possibly aligned in a way not
honouring its .sh_addralign.

We arrange for this alignment raising by setting the aligment of the
section's data buffer to match the containing segment's sg_align --
which must be sufficient to align the section with the greatest
requirement -- possibly overstating what's actually necessary.

This has the potential to be both grossly incorrect and grossly
destructive.  It has only been very very lightly tested.