ld: Actually align based on the first non-empty section
Previously, we were padding out to the most-aligned section in a
segment, for convenience.  This is provably safe, but also wasteful if
the first non-empty section requires a lesser alignment.  Instead, walk
the data descriptors of the first non-empty section, and pad to _its_
alignment.

We have to walk the descriptors ourselves, because the sections
sh_addralign hasn't been set by libelf`elf_update yet.