libc: Don't declare SPARC __rtld static
The libc-as-interpretter bootstrap is built in a rather odd way, we have a bit
of stock assembler, _rtboot.s, and some C, _rtld.c.  We build them by using cc
-S on _rtld.c, appending that to _rtboot.s and assembling the result.  This
means that __rtld _can_ be static, because it's built in the same unit as its
caller, _rtboot.s.  However, the compiler has utterly no way to know this, and
a compiler which choose to not emit unused static functions will thus not emit
__rtld.

Rather than telling GCC that while the static __rtld is unused, we need it
emitted just cease declaring it static (or .local), as we don't on i386.
There is no symbol collision problem here, as it's scoped local by the libc
mapfile anyway, and so invisible to callers, and we're the only definition
within libc.