Mark functions which return unusually (twice, skip a frame, etc.)
GCC has innate knowledge that certain functions return twice (vfork, setjmp,
getcontext).  Unfortunately, we have a handful of others (on_trap, on_fault,
some in the deep innards of libc stack unwinding), and we have vforkx which is
the innards of vfork, but also of posix_spawn, etc.

This is necessary because, without the hint, the compiler may think a variable
is dead because it is unreferenced because we have passed the last variable
reference after the _first_ return, and clobber a variable, only to leave us
consuming trash after the _second_ return.

This, in particular, has been true of posix_spawnp on SPARC.