saveargs: only search for prologue in insns which likely executed
Limit our search for the argument saving instructions to those prior to
the program counter, such that we don't display junk.

It is in theory possible for this to misfire if, for eg, the argument
saving code is run and then we take a backward jump, but in practice
this is very unlikely, and branches prior to that part of the prologue
would be a separate -- and gross -- cause for concern.