dtrace: add runtime support for objects that can't load their own DOF
The regular D runtime is linked into each ELF container object by dtrace -G,
and bound hidden.  This means that each each object contains support for
loading itself at runtime regardless of how many such objects are
involved in a link -- if there are 2 providers, there are 2 copies of
the runtime.

We need support for ELF containers that cannot load their own DOF, but
require a single external runtime to load all of their DOF (from however
many providers, etc, exist).

We do this by placing a symbol __SUNW_dof_array into the output object,
which is a NULL-terminated array of pointers to all DOF which must be
loaded by the new runtime object (ldrti.o), as well as libdtrace and
libdtrace_forceload.

The __SUNW_dof_array symbol is placed into a new .SUNW_dof_array section
of type SHT_SUNW_DOF_ARRAY, similar to the .initarray and .finiarray
sections in concept though this is an implementation detail (and one
that it's hoped will change).