Add limit sets to security-flags, zone configuration of security-flags
Add 'upper' and 'lower' sets to security-flags, specifying the maximal
and minimal allowable flags, respectively.  No set is allowed to exceed
these bounds, including the sets themselves (which means that
restriction may only ever increase).

Add a 'security-flags' resource to zones, to configure the set of flags
applied to that zone.

Change the svc:/system/process-security properties to allow
specification of upper/lower/default (what used to be security_flags is
now default).