tty: up the input queue limit to 2K
This heavy handedly ups the limit for TTY input to 2 kilobytes.

This includes for the line-editing provided by canonical mode
(previously 256 bytes), and input in genereal (previously 512 bytes).

This is heavy handed because I just upped the queue lengths and high
water marks all the way up and down the stack both in the modules, and
the values we push up to the stream head.

I have left the constants that "defined" (they didn't, they just matched
it) these limits alone, as that seems to be what both FreeBSD and Linux
have done, and I'm not at all sure what the implications of being
different would be.

These limits are
CANBSIZ
MAX_CANON
MAX_INPUT
_POSIX_MAX_INPUT
_POSIX_MAX_CANON

I added a new symbol with the _real_ value, which hopefully makes things
easier to follow.

I've tested this with cat | wc -c, and get the expected value of 2049
(because of the newline).