libc: fully initialize mutex_t when initializing POSIX robust mutexes
Sun mutexes demand that a mutex be zeroed completely _prior_ to
mutex_init() in the case of robust mutexes, and that we return EBUSY if
mutex_init() is called twice on the same mutex.

pthread_mutex_init() has no such restriction.  If pthread_mutex_init is
handed a robust mutex without the mutex magic, fully 0 it before passing
it on to mutex_init().