XXXXX hidden undefined weak symbols should not leave relocations
If a hidden weak symbol is undefined, it can never become defined after
the link edit completes because of the now-local binding.  Do not leave
behind a relocation for it.

Previously we would leave behind symbolic relocations with no associated
symbol, because the reduction would remove the symbol from the symbol
tables.  Such relocations are invalid, and would fail at runtime, so
similarly expand the checking for such relocations on output to catch
the case where the associated symbol has been reduced away (which occurs
after the current check).