man-fixer: A script to fix up normative manual page references, where possible
Update references in place if we're confident we see a manual page
reference (the page to which it refers exists in the section to which it
refers in two big distributions, prior to our renames).  If we're not
sure, log a message with file/line/page so for manual fixes.