Disable conditional dependencies on python modules
The build of the in-gate python modules (pysolaris, pylibbe, pyzfs)
is currently disabled, but we do have a working python-311 package
in the OmniOS repository. When this is installed, system/library suddenly
gains a dependency on solaris-311, and zfs gains one on pyzfs-311.
While the python modules are disabled, so should the conditional dependencies
be.