libpctx: stop shipping (and fix clean build)
This is part of cpc, which isn't yet ported, more importantly it requires
libcpc (actually, only libcpc.h) which we don't build, and only accidentally
had in the proto area.

Removing this broke the libpctx build, but that was covered by gcc having a
stashed copy of the target headers (frustratingly).

Fix the clean build, by not shipping this either.