ldd: Fix platform detection and implement lddstub
This fixes yet another piece where we mis-detect the platform as 32bit,
because the way we do this is to compare the platform string, and know whether
that string means 32 or 64bit.  Eugh.

This also implements lddstub, a simple dynamic executable that may not use any
shared libraries, and just calls exit(2), used for ldd(1) to do it works on
shared libraries.