fm: Add a basic featureless legacy-style enumerator so we have a topology
This just adds a topo equivalent to the x86 legacy topo, empty except for a
mainboard.  This gives us _a_ topology, and calms down problems in the fault
manager.

It also because of the topology search order still leaves us room to do more
specific things for individual implementations (uname -i), or if SBSA is a
different machine (uname -m).