uts: Fully wire up `poke_cpu`
On systems with > 2 cores we would hang during boot as each cpu after the 2nd
came online and waited for the CPUs to pause.  An interrupt (via the console
for eg) would shake this loose for a cpu, and then we'd go back and do it
again.

This turns out to be because the already online CPUs other than the boot CPU
enter the dispatcher and start running the idle thread, which ultimately waits
for interrupts in a `wfi`.  However we never got out of this state in the
normal case, because `poke_cpu` didn't actually do anything, and so we never
received an SGI when `cpu_wakeup` poked the CPU to be woken.  This means that
when `cpu_pause_start` set the CPU to be paused and called `setbackdq`,
nothing actually happened to the target CPU.

Fix this by implementing the bits of CPU poking that seem to have been
intended but were not present, so that we receive an SGI `IRQ_IPI_CPUPOKE` at
`XC_CPUPOKE_PIL` and handle it just by acknowledging the interrupt but doing
nothing else.

Two implications of this are:

1) It was remarkably unlucky to have picked a qemu setup by default that only
   had 2 cores and was immune to this.

2) After booting a sufficient amount to avoid this same basic issue, the
   system's scheduling, etc, was working effectively by chance as we received
   interrupts at a sufficient rate to make things appear fine.