aarch64: clean up EFI boot support
The basic support for this was already present but it required
the zpool to be in an EFI partition labelled as EFI_ROOT whereas
ZFS uses EFI_USR for pools, and several places within libzfs
and libzpool expect this.

Restyled the code to illumos style, added more prom prints and
allowed EFI_USR partitions too.

With this, an EFI disk image can boot on both qemu and an rpi4,
which is nice because it is then possible to expand the ZFS
pool to fill the disk.