aarch64: enable hsfs boot archives in altroots too
When doing a pkg update into an alternate BE, the boot archive
in that BE is updated. However, an HSFS archive is not currently
allowed there unless the bootadm command is run directly. Moving
the aarch64 escape hatch up slightly resolves this and allows
the boot archive to be updated automatically.