don't forcibly load procfs twice.
We load procfs (and certain others) in `startup_modules()`, but also in
`post_startup()`, with a truly ancient comment saying to stop doing it in
`post_startup()` so let's.

There are also stubs for procfs, despite us forcibly loading it, and it not
being unloadable.  I'm leaving those alone as I can't work out how they ever
became necessary.

This is also fixed on the intel side to avoid letting that slip.