Add "name" and "unit-address" props to the FDT, to more closely match 1275
According to the devicetree specification[1] what in 1275 would be the "name"
and "unit-address" properties of a prom node are instead reflected in the
inherent name of the node (which 1275 does not have).

Rather than trying to fake these properties in `prom_getprop`, `prom_setprop`,
etc. just add these properties to the tree as we build it.

[1]: https://devicetree-specification.readthedocs.io/en/v0.3/devicetree-basics.html#properties