tests/os-tests/cores: don't try to run 32-bit tests on aarch64
The core tests expect a 32bit and 64bit process for their checks, but on
aarch64 we only have 64bit.  Adapt to that, and only try check for and run the
32bit binary on other platforms.

Unfortunately, this is in terms of machine name, because it appears there's no
way to get out of isainfo(1) the full set of bit-ness we support.  (`isainfo
-b` says "64" not "64 32")