ipsecconf: use memmove(3C) not strlcpy(3C) where buffers may overlap
We can't use strlcpy(3C) to fill `lo_buf`, as we may be reading from
`*leftover` which is a pointer into `lo_buf`, this gives us a good
chance of copy being between overlapping buffers, which is undefined
behaviour.

Calculate the length ourselves, and use memmove(3C) which supports
overlapping buffers.