overlay(4D): allow 64bit processes to open the clone device
When a 64bit process does open(2) the kernel inserts `FOFFMAX` into the flags
to indicate to the rest of the system that this handle is largefile aware, the
same as if a 32bit process did open64(2).

`overlay_target_open` was too restrictive in forbidding invalid flags and
would forbid that one as well, returning `EINVAL` and effectively forbidding a
largefile-aware or 64bit process from working.