float.h: adjust macro values for aarch64
These values were expanded from the compiler builtins and
verified against values documented in the <float.h> table of
the Arm Compiler for Embedded Reference Guide for accuracy.

Note that the compiler guide documents DECIMAL_DIG with the
value of 17, however the compiler builtin reports this as 36.

Linux also reports this as 36, but FreeBSD and (arm64) macOS reports
this as 17. Since the compiler guide uses 17 I figured that it would
make sense to err on the side of caution and use 17 for now.