remove `_SYSCALL32` from AArch64
This fix is largely a matter of correctly adding `_SYSCALL32_IMPL` guards to
32bit compatibility code that was missing them, thereby removing that code
from 64-bit only AArch64.  With those sections guarded, the need for (almost
all) the types goes away too.

Some things had to be moved from under `_SYSCALL32` guards to facilitate this
as over the years their use became generic as there was no platform on which
`_SYSCALL32` wasn't defined to catch this.

The larger turmoil is oddly in userland where (the UCB side of) ps(1), and
truss(1) were particularly unclean.  Guards have been added and limited
amounts of code adjusted to cope.