simple-bus: merge the dangling meson-gxbb simple-bus driver into the real one
For reasons I'm not sure of, meson-gxbb had its own simple-bus with better (I
think) support for mapping.  Merge the two together (really, take the meson
one) and fix some minor cstyle nits (but not the major ones).

It's obvious this could use some more attention, but this seems a good start.