libproc: syscalling on aarch64 must explicitly use indirect system calls
The implementation of `Pissyscall()` uses an existing system call instruction
in the program text to which it can jump to enter the kernel with a crafted
set of arguments.

Unfortunately, when searching for an instruction via `Pissyscall_text()` or
using one we know of via `P->sysaddr` and checking it via `Pissyscall()` and
`Pissyscall_prev()` we search for the location of any system call instruction
as we would on intel or sparc.  Unfortunately AArch64 system calls pass the
system call number encoded in the instruction, so this is very likely to land
upon a direct system call -- which will ignore the system call number in
`%x9`, and perform an arbitrary system call instead using crafted arguments.

In the best case, this would fail.

Fix this by checking explicitly for an indirect system call (`svc #0`) in
`Pissyscall_text()`, and adding private analogues of `Pissyscall() and
`Pissyscall_prev()` so that `Pscantext()` and sysaddr caching can take the
same precaution.