add a top-level toggle indicating the absence of java
`ENABLE_JAVA` can be set to `$(POUND_SIGN)` to indicate that Java is entirely
absent, and should not be used.  It defaults this way on AArch64.

add `JAVA_ONLY` to pkg processing, to guard those files that should only be
delivered when java is present.  Add an `org.illumos.java-only` attribute
which indicates entire packages should only exist when java is present.