pkg: Introduce org.illumos.for-arch
This should now be used in place of `variant.arch`, and `variant.arch`
allowed to be derived automatically.

This is necessary because the overloaded of `variant.arch` for this purpose
cannot work to specify multiple architectures, as it also causes the files to
be delivered for both architectures (that is, each file twice, both
architectures on each architecture).  We currently work around this by
re-setting `variant.arch` to always refer to the current architecture alone.

Deal with this more fully by using a new, self-deleteing property
`org.illumos.for-arch` to control whether a package is included on a platform,
and automatically insert `variant.arch` for the current arch, rather than
giving `variant.arch` two subtly incompatible meanings and then fixing them
up.