bc: port and connect to the build
The problem here is that bc(1) uses `int` to store either pointers or integers
in "bundles", which is a mess when 64bit.

Fix this by storing `intptr_t` which can at least store both.

A better long-term fix here would be to port Otto Moerbeek's rewrite for
OpenBSD, which is much better, but would require compatibility
investigation both with historic behaviour and with our dc(1).

I have tested this by calculating π to 998 places and comparing both the
result and the generated `dc(1)` input with x86, which requires the arc
tangent implementation in `lib.b` and is the best I can come up with to
exercise bundles, etc.  Perhaps a similar -- but more thorough -- means can
test the OpenBSD implementation.