aarch64: implement arch_timer abstraction
This abstraction allows the kernel to select the physical or
virtual counter depending on the boot EL. If the kernel is booted
at EL2, then the physical counter is used. If the kernel is booted at EL1,
then the virtual counter is used. This frees up the physical counter
for use by a hypervisor if booted at EL2, and prevents traps to EL2
if the kernel is booted at EL1 (from within a hypervisor or otherwise).

For now the bootloader is fixed to the physical counter regardless
of EL (this mirrors behavior seen in other bootloaders like u-boot).