libc: preserve the TCB when recycling ulwp_t on variant 1 TLS platforms
Previously, when we recycled a thread off the dead list, we'd zero the whole
`ulwp_t` except for the stack information.  On variant 1 TLS platforms we
absolutely must preserve the `tcb_t` too, as it's where the thread-pointer
points.

Failure to do this means that `__curthread()` (two-underscores) will return
NULL, while `_curthread()` (one-underscore) will continue to return our
`ulwp_t` (see the difference between the two described in thr_inlines.h).

There are as you may imagine varied symptoms to this.  The primary and most
noticeable happens due to an optimization in `lmalloc()`.  If `__curthread()`
is `NULL` lmalloc uses the global `&__uberdata` but also _does not do any
locking_ on the assumption that there is only one thread.  Because of the
cleared TCB of any recycled thread this means that any recycled thread did no
locking during internal to libc allocation.

The most common, but difficult, symptom of this is that two threads, in
`lmalloc()` for a small allocation at the same time, could receive the same
pointer in return.  Further symptoms include freelist corruption, other
symptoms of the inherent use-after-free when one thread frees
this memory, and double free when the other does as well.

The most repeatable symptom of this is that libc-test's `wcsrtombs()` test
would fail in the threaded case repeatably, but after a variable amount of
time and in variable ways.  The rapid cycling through threads in this test
practically guaranteeing that we'd recycle them often, and lose one of the
races.  Other victims of this same basic issue would be atexit(3C) and
related, error reporting in the case of assertion failures or libc panicks,
and the failure of thr_main(3C).

It is possible, but not confirmed, that the occasional bad `lfree()`
from svc.configd(8) is this same bug.