libc: doprnt needs to know `char` might be unsigned
When printing a negative character using `%hhd` printf(3C) and friends would
cast the value to `char` in `_ndoprnt()`, and lose the sign.  Cast to `int8_t`
instead, which is what we actually mean.  This causes no change in the x86
binaries.

This fixes the failure of the libnvpair json tests in util-test.