libc: `__mul_set` must ensure the multiplication happens before we check flags
In `__mul_set` we perform double precision multiplication, and then check the
FP flags for exceptions.

Unfortunately, the result of the multiplication was not declared volatile and
`read_fpsr`/`read_fpcr` have no way to say the condition codes are inputs, so
the compiler could (and did) move initialization to after the flag checks,
making the flag checks invalid.  This meant that `__fast_double_to_decimal`
would always give the impression it succeeded even with FP traps, and is the
root cause of the incorrect floating point formatting of certain values with
certain numbers of significant figures.

Declare it, and the new status, `volatile`, so that assignments to them cannot
be reordered across sequence points by the compiler.