switch to our own exception vector much earlier
Previously we would stay on the exception vector from the loader until
`startup()` was complete, which is a desperately long time as these things go.
This probably risks it actually disappearing under us, to at least a slight
degree.

More importantly it means we don't get real panicks, and gets in the way of
other things that need traps early.

Move the switch to the earliest possible point we can handle our own traps --
just after we have a system timer.