getzoneid: return `GLOBAL_ZONEID` rather than crashing in early boot
If `curproc->p_zone` is `NULL` then we're in early boot before zones is initialized,
and `curproc` is `p0`.  In this case we can only be the global zone.

I've done this so that panics in early boot don't recursively panic attempting
to show the system dump messages on the user console which calls
`getzoneid` in `cprintf` when `SL_USER`, this makes life easier for debugging,
and I don't believe there is another valid case where `p_zone == NULL`, but
I'm not 100% on the correctness of this.