Add a cpuinfo helper
Add a cpuinfo helper, based on the ACPI 6.5 GIC CPU Interface table structure
as found in the ACPI MADT.  The idea behind this structure is modelled on a
similarly named structure in NetBSD.

The goal of this structure is to make MP startup and ncpus calculation firmware
independent, so wire up the new (FDT-only) code there.

In the i86pc code this approach (looking at the MADT) is used to determine the
number of CPUs in the machine, so this is not a novel approach by any means.