Add a psciinfo helper
Add a helper for gathering PSCI information from firmware, then tweak the PSCI
code to use the new helper and initialise much earlier in the boot process,
allowing early reboot to work as expected.

Similarly to the cpuinfo change, the goal of this work is to make startup
firmware independent.