Reintroduce sys/prom_debug.h
In an effort to reduce the unnecessesary differences between i86pc and armv8,
reintroduce prom_debug.h as a place to store startup debugging macros.

There's no functional change in this diff - the 32bit vs 64bit printing
format lookup is taken care of with a cast in the i86pc implementation.