Don't rely on FDT for boot properties
To make the startup code reusable between devicetree and ACPI we
need to reduce our dependency on FDT for communicating boot properties
between inteboot (or any loader shim) and the kernel boot code.

Use a combination of xboot_info and boot modules to disentangle this
part of the codebase.

Since the illumos FDT integration requires a phandle for all nodes,
and libfdt only insists on phandle for referenced nodes, have the
`inetboot` shims tweak the FDT before passing it through to illumos.

Also have the `inetboot` shims ensure that well known pseudo-nodes exist
before starting `unix`.

While I'm here, make all files touched pbchk clean.

This diff introduces a few new XXXARM blocks, which are in place to
track places we need to adjust to get parity with i86pc.