Configure PSCI from xboot_info
We have all the accurate information we need to configure PSCI in the
bootloader/shim. Communicate PSCI configuration to `unix` via the
`xboot_info` structure and take advantage of this information so that we
can configure PSCI much earlier in the boot.

While here, clean up pbchk in the files I've touched.