Turn on -msave-args on aarch64
This means that, like, amd64, we save integer arguments on the stack on
function entry so the debugger etc can find them again statically.  This is
the part that enables the compiler switch and teaches libsaveargs, mdb and
libproc about what has happened.

You need the matching compiler changes which implement the flag.