os-tests/tests/stackalign: port
This tests that we have sufficient stack alignment in a bunch of common places
it might go wrong.  Luckily, it appears we do.

This renames the 32bit source files such that we can refer to them with
`$(MACH32)`, because on aarch64 `$(MACH)` and `$(MACH64)` are the same.  Where
they're built in the same directory, we have no alternative but to
disambiguate.