tools: put which_scm executably into the workspace
which_scm needs to be runnable out of the workspace, if we want to build on
systems that don't have their own onbld tools _at all_ like while
bootstrapping arm64.

This makes which_scm work like nightly and bldenv, where they're directly in
the workspace.  This means when `nightly` looks for which_scm in whatever
directory it itself is in, for preference, it finds the copy also in
`usr/src/tools/scripts`