Widen regspec and rangespec
Widen the fields in both regspec and rangespec to be 64-bit on aarch64.

This change allows us to use 64-bit addresses in the same way as
regspec64 (regspec is now identical to regspec64) and allows us to use
both 64-bit addresses and 64-bit sizes in ranges, gracefully handling
PCIe range sizes >4GiB.

This also allows us to simplify simple-bus and viommionex, having them
use the implementation DDI for bus mapping and interrupt operations.