loader: Add acpica for loader on aarch64
This change provides acpica compiled for the loader environment. There
are three parts to this change:
* Simple Makefile support
* Implementation of mandatory integration functions, based closely on
  those found in edk2, but cleaned up to avoid undersirable behaviour
  and properly handle hardware reduced platforms. This includes removing
  support for legacy ioports, which don't exist on aarch64.
* A simplified set of initialisation and finalisation functions to be
  used around usage of ACPI.

Aside from the init/fini functions, the most interesting part of this
change is that the default integration functions would call UEFI
Runtime Services to get the system time. On aarch64 systems one can't
call UEFI Runtime Services until you are at your final exception level,
and since we don't know what that will be at this stage in the boot, we
have reimplemented the `AcpiOsGetTimer` function in terms of the
architected timer. This function is called as part of avoiding infinite
loops when interpreting AML, so it will always be called during
namespace initialisation and may be called while interrogating the
system bus namespace.