loader: Fix some declaration clashes for aarch64
The loader code contains two competing implementations of C library
functionality (those that live in standard-like headers and those that
live in stand.h). These declarations are often incompatible, which
creates challenges when mixing code that needs one style with code that
needs the other (in this case, acpica and loader code).

Add a few conditional tweaks to avoid the undesirable (and in some cases
unimplemented) parts of the FreeBSD libc that could otherwise be
referenced in a loader build on aarch64.