loader: Add an FDT utilities library
The FDT utilities library creates, on-demand, a read/write copy of the
DTB blob provided by UEFI. Prior to providing the blob to a caller,
the devicetree is normalised to include nodes expected by illumos and
all nodes are updated to include a phandle, which is required by the
illumos device tree integration.

The FDT blob is only ever returned when the booted system does not
offer ACPI tables. The Arm EBBR states that when both firmware table
implementations are present that only ACPI should be offered, so this
is just an extra enforcement point that should be unnecessary if the
bootloader is compliant.

The library also includes code for deriving platform details, much
the same way as `inetboot` has done in the past.

Both changes are hooked into loader in this diff.